Refactored server code
