+++ .gitignore
@@ -1,4 +1,4 @@
 /node_modules
-/env/.env.dev
-/env/.env.test
+.env.dev
+.env.test
 *.log
