+++ server/schema/resolvers/message.js
@@ -1,6 +1,6 @@
-const { formatErrors } = require('../utils');
-const models = require('../models');
-const pubsub = require('../pubsub');
+const { formatErrors } = require('../../utils');
+const models = require('../../models');
+const pubsub = require('../../pubsub');
 
 module.exports = {
   Query: {
