+++ server/schema/types/friend.js
